image_recognition_project/requirements.txt
Pillow>=8.0.0
requests>=2.25.1
Flask>=2.0.1
Werkzeug>=2.0.1
gunicorn>=20.1.0
google-generativeai>=0.3.0
